build adjustments
When defining a functional component the way we have with the Button element, we must also provide a displayName as one is not automatically generated via the function name as there is no IIFE to grab the name from.

Unsure of the reason why testing-library-react doesn't like passing children to layout as props, but have fixed that as well by wrapping the child element in the layout instead of passing the child as a prop.